Update game from prompt-based to button based
I've updated the RPS game to allow the player to click buttons to pick
their choice. Round results and game results now print to the page instead
of just the console!

The main thing that probably should be added is a button to restart the
game...

Lots can still be done here; there's no CSS file (yet...?) for example,
to make everything look nice. But the basic functionality is here.